fix label bounding box bug

+++ MOT_Thesis_1551025_1551034/BoundingBox.cpp
@@ -23,7 +23,7 @@ int BoundingBox::getClassId()
 	return classId;
 }
 
-int BoundingBox::getConfidence()
+float BoundingBox::getConfidence()
 {
 	return confidence;
 }
